Move all urls into tests/urls.py.
